krijg de sleep libarary nog niet helemaal werkend

+++ CSD2A/python_basics/opdracht1/rangeloop_test.py
@@ -1,5 +1,5 @@
 user_input = input('Enter space-separated integers: ').split()
 
-list_of_integers = [int(item) for item in user_input]
+list_of_integers = [float(item) for item in user_input]
 
 print(list_of_integers)
